requirements.txt
flask
flask-mysql
flask-bcrypt
flask-session
mysql-connector-python